requirements.txt
streamlit
opencv-python
ultralytics
torch
numpy
Pillow
PyYAML